+++ libs/docs/core/dynamic-page/index.ts
@@ -1,6 +1,7 @@
 import { Routes } from '@angular/router';
 import { DynamicPageDocsComponent } from './dynamic-page-docs.component';
 import { DynamicPageDocsHeaderComponent } from './dynamic-page-header/dynamic-page-docs-header.component';
+import { DummyComponent } from './examples/dynamic-page-routing/dynamic-page-routing-example.component';
 
 export const ROUTES: Routes = [
     {
@@ -9,7 +10,30 @@ export const ROUTES: Routes = [
         children: [
             {
                 path: '',
-                component: DynamicPageDocsComponent
+                component: DynamicPageDocsComponent,
+                children: [
+                    {
+                        path: 'example1',
+                        component: DummyComponent,
+                        data: {
+                            text: `I'm example 1`
+                        }
+                    },
+                    {
+                        path: 'example2',
+                        component: DummyComponent,
+                        data: {
+                            text: `I'm example 2`
+                        }
+                    },
+                    {
+                        path: 'example3',
+                        component: DummyComponent,
+                        data: {
+                            text: `I'm example 3`
+                        }
+                    }
+                ]
             }
         ],
         data: {