+++ libs/platform/icon-tab-bar/components/icon-tab-bar-text-type/icon-tab-bar-text-type.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { IconTabBarComponent } from '../../icon-tab-bar.component';
-import { generateTestConfig } from '../../tests-helper';
+import { _generateTabBarItems, generateTestConfig } from '../../tests-helper';
 import { IconTabBarTextTypeComponent } from './icon-tab-bar-text-type.component';
 
 describe('IconTabBarTextTypeComponent', () => {
@@ -18,7 +18,7 @@ describe('IconTabBarTextTypeComponent', () => {
     beforeEach(() => {
         fixture = TestBed.createComponent(IconTabBarTextTypeComponent);
         component = fixture.componentInstance;
-        component.tabsConfig = generateTestConfig(10);
+        component.tabs = _generateTabBarItems(generateTestConfig(10));
         fixture.detectChanges();
     });
 
@@ -27,17 +27,17 @@ describe('IconTabBarTextTypeComponent', () => {
     });
 
     it('should _selectExtraItem move selected tab to last visible position', () => {
-        const hiddenItem = component._tabs$()[component._tabs$().length - 1];
+        const hiddenItem = component.tabs[component.tabs.length - 1];
 
         component._selectExtraItem(hiddenItem);
-        const lastVisibleItemAfterSelect = component._tabs$()[component._lastVisibleTabIndex];
+        const lastVisibleItemAfterSelect = component.tabs[component._lastVisibleTabIndex];
 
         expect(hiddenItem.label).toEqual(lastVisibleItemAfterSelect.label);
     });
 
     it('should insert tab to another tab as child', () => {
-        const draggableItem = component._tabs$()[0];
-        const targetItem = component._tabs$()[1];
+        const draggableItem = component.tabs[0];
+        const targetItem = component.tabs[1];
         component._onDropped({
             draggableItem,
             targetItem,
@@ -48,8 +48,8 @@ describe('IconTabBarTextTypeComponent', () => {
     });
 
     it('should replace tab', () => {
-        const draggableItem = component._tabs$()[0];
-        const targetItem = component._tabs$()[1];
+        const draggableItem = component.tabs[0];
+        const targetItem = component.tabs[1];
         const dragItemPreviousUid = draggableItem.uId;
         const targetItemPreviousUid = targetItem.uId;
 
@@ -65,10 +65,10 @@ describe('IconTabBarTextTypeComponent', () => {
     });
 
     it('should change number of visible/hidden tabs', () => {
-        const tabsLength = component._tabs$().length;
+        const tabsLength = component.tabs.length;
         const extraTabs = 5;
         component._recalculateVisibleItems(extraTabs);
-        const visibleTabsLength = component._tabs$().filter((tab) => !tab.hidden).length;
+        const visibleTabsLength = component.tabs.filter((tab) => !tab.hidden).length;
 
         expect(component._extraTabs$().length).toBe(extraTabs);
         expect(visibleTabsLength).toBe(tabsLength - extraTabs);