+++ libs/platform/icon-tab-bar/icon-tab-bar.component.ts
@@ -1,26 +1,47 @@
 import { NgTemplateOutlet } from '@angular/common';
 import {
+    AfterViewInit,
     ChangeDetectionStrategy,
     ChangeDetectorRef,
     Component,
+    ContentChildren,
     DestroyRef,
+    ElementRef,
     EventEmitter,
+    HostBinding,
     Input,
     OnInit,
     Optional,
     Output,
+    QueryList,
+    TemplateRef,
+    ViewChild,
+    ViewChildren,
     ViewEncapsulation,
+    booleanAttribute,
     computed,
-    inject
+    inject,
+    signal
 } from '@angular/core';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
-import { ContentDensityService, RtlService } from '@fundamental-ngx/cdk/utils';
+import { ContentDensityService, Nullable, RtlService, scrollTop } from '@fundamental-ngx/cdk/utils';
+import { FD_DYNAMIC_PAGE } from '@fundamental-ngx/core/dynamic-page';
 import { IconFont } from '@fundamental-ngx/core/icon';
-import { distinctUntilChanged } from 'rxjs/operators';
+import { ScrollSpyDirective } from '@fundamental-ngx/core/scroll-spy';
+import { ScrollbarDirective } from '@fundamental-ngx/core/scrollbar';
+import { FD_TABLIST, TabList } from '@fundamental-ngx/core/shared';
+import { fromEvent } from 'rxjs';
+import { debounceTime, distinctUntilChanged, first, startWith } from 'rxjs/operators';
+import { IconTabBarBase } from './components';
 import { IconTabBarFilterTypeComponent } from './components/icon-tab-bar-filter-type/icon-tab-bar-filter-type.component';
 import { IconTabBarIconTypeComponent } from './components/icon-tab-bar-icon-type/icon-tab-bar-icon-type.component';
 import { IconTabBarProcessTypeComponent } from './components/icon-tab-bar-process-type/icon-tab-bar-process-type.component';
+import {
+    IconTabBarTabComponent,
+    IconTabBarTabContentDirective
+} from './components/icon-tab-bar-tab/icon-tab-bar-tab.component';
 import { IconTabBarTextTypeComponent } from './components/icon-tab-bar-text-type/icon-tab-bar-text-type.component';
+import { UNIQUE_KEY_SEPARATOR } from './constants';
 import { IconTabBarItem } from './interfaces/icon-tab-bar-item.interface';
 import { TabColorAssociations } from './interfaces/tab-color-associations.interface';
 import { TabConfig } from './interfaces/tab-config.interface';
@@ -33,15 +54,24 @@ import { IconTabBarBackground, IconTabBarSize, TabDestinyMode, TabType } from '.
     changeDetection: ChangeDetectionStrategy.OnPush,
     encapsulation: ViewEncapsulation.None,
     standalone: true,
+    providers: [
+        {
+            provide: FD_TABLIST,
+            useExisting: IconTabBarComponent
+        }
+    ],
     imports: [
         NgTemplateOutlet,
         IconTabBarProcessTypeComponent,
         IconTabBarFilterTypeComponent,
         IconTabBarIconTypeComponent,
-        IconTabBarTextTypeComponent
+        IconTabBarTextTypeComponent,
+        ScrollbarDirective,
+        ScrollSpyDirective,
+        IconTabBarTabContentDirective
     ]
 })
-export class IconTabBarComponent implements OnInit {
+export class IconTabBarComponent implements OnInit, AfterViewInit, TabList {
     /**
      * @description Type of tab bar view.
      */
@@ -52,7 +82,13 @@ export class IconTabBarComponent implements OnInit {
      * @description A tab bar configuration that stores the state of each tab. Based on this configuration, a tab bar is representing.
      */
     @Input()
-    tabsConfig: TabConfig[] = [];
+    set tabsConfig(value: TabConfig[]) {
+        this._tabsConfig$.set(value);
+    }
+
+    get tabsConfig(): TabConfig[] {
+        return this._tabsConfig$();
+    }
 
     /**
      * @description Destiny mode.
@@ -104,6 +140,20 @@ export class IconTabBarComponent implements OnInit {
     colorAssociations: TabColorAssociations;
 
     /**
+     * Whether to open tab content one under another without collapsing.
+     * Works only for content-projected tab content.
+     */
+    @Input({ transform: booleanAttribute })
+    stackContent = false;
+
+    /**
+     * Maximum height of the content.
+     * Works only for content-projected tab content.
+     */
+    @Input()
+    maxContentHeight = '100%';
+
+    /**
      * @description Emits when some tab is selected.
      */
     @Output()
@@ -120,13 +170,73 @@ export class IconTabBarComponent implements OnInit {
     closeTab = new EventEmitter<IconTabBarItem>();
 
     /** @hidden */
+    @ContentChildren(IconTabBarTabComponent, { descendants: false })
+    children: QueryList<IconTabBarTabComponent>;
+
+    /** @hidden */
+    @ContentChildren(IconTabBarTabComponent, { descendants: true })
+    allChildren: QueryList<IconTabBarTabComponent>;
+
+    /** @hidden */
+    @ViewChild(ScrollbarDirective)
+    _scrollbar: ScrollbarDirective;
+
+    /** @hidden */
+    @ViewChildren(IconTabBarTabContentDirective)
+    tabDirectives: QueryList<IconTabBarTabContentDirective>;
+
+    @HostBinding('class.fd-tabs-custom')
+    private get _customTabs(): boolean {
+        return this._inDynamicPahe;
+    }
+
+    @ViewChild(IconTabBarBase)
+    private readonly _iconTabBarCmp: Nullable<IconTabBarBase>;
+
+    /** @hidden */
+    get headerContainer(): Nullable<ElementRef> {
+        return this._iconTabBarCmp?.headerElement;
+    }
+
+    /** @hidden */
     _cssClassForContainer: string[];
+
+    /** @hidden */
+    _disableScrollSpy = false;
+
+    /** @hidden */
+    _init = true;
+
+    /** Scrollable element reference. */
+    get scrollableElement(): Nullable<ElementRef> {
+        return this._scrollbar?.elementRef;
+    }
+
+    /** @hidden */
+    readonly _selectedUid = signal<string | undefined>(undefined);
+
+    /** @hidden */
+    readonly _tabs$ = computed(() => this._generateTabBarItems(this._tabsConfig$()));
+
+    /** @hidden */
+    readonly _flatTabs$ = computed(() => this._generateFlatTabs(this._tabs$()));
+
+    /** @hidden */
+    readonly _tabRenderer$ = signal<IconTabBarItem | null>(null);
+
+    /** @hidden */
+    readonly _tabRenderers$ = signal<{ renderer: TemplateRef<any>; id: string }[]>([]);
 
     /** @hidden */
     readonly _rtl$ = computed(() => !!this._rtlService?.rtlSignal());
 
     /** @hidden */
-    private _destroyRef = inject(DestroyRef);
+    readonly _tabsConfig$ = signal<TabConfig[]>([]);
+
+    /** @hidden */
+    private readonly _destroyRef = inject(DestroyRef);
+
+    private readonly _inDynamicPahe = !!inject(FD_DYNAMIC_PAGE, { optional: true });
 
     /** @hidden */
     constructor(
@@ -156,6 +266,29 @@ export class IconTabBarComponent implements OnInit {
         }
     }
 
+    /** @hidden */
+    ngAfterViewInit(): void {
+        this.children.changes.pipe(startWith(null), takeUntilDestroyed(this._destroyRef)).subscribe(() => {
+            const projectedTabs = this.children.toArray();
+
+            if (projectedTabs.length === 0) {
+                return;
+            }
+
+            this.tabsConfig = projectedTabs.map((t) => this._generateTabConfig(t));
+            this._cd.detectChanges();
+        });
+
+        this.allChildren.changes.pipe(startWith(null), takeUntilDestroyed(this._destroyRef)).subscribe(() => {
+            this._tabRenderers$.set(
+                this.allChildren.map((c) => ({
+                    renderer: c.renderer,
+                    id: c.id
+                }))
+            );
+        });
+    }
+
     /**
      * @hidden
      * @param event reordered array of IconTabBarItem
@@ -169,12 +302,39 @@ export class IconTabBarComponent implements OnInit {
      * @param selectedItem
      */
     _selectItem(selectedItem: IconTabBarItem): void {
+        this._tabRenderer$.set(selectedItem);
         this.iconTabSelected.emit(selectedItem);
+
+        if (this.stackContent) {
+            this._scrollToPanel(this.tabDirectives.find((tab) => tab.uId === selectedItem.uId)!);
+        }
     }
 
     /** @hidden */
     _closeTab(item: IconTabBarItem): void {
         this.closeTab.emit(item);
+    }
+
+    /** Programmatically select tab. */
+    selectTab(id: Nullable<string>): void {
+        const selectedItem = this._flatTabs$().find((t) => t.id === id);
+        if (!selectedItem) {
+            return;
+        }
+        this._selectedUid.set(selectedItem.uId);
+        this._selectItem(selectedItem);
+    }
+
+    /** @hidden */
+    highlightActiveTab(id: Nullable<string>, scroll = false): void {
+        if (this._disableScrollSpy) {
+            return;
+        }
+        const activeTab = this.tabDirectives.find((tab) => tab.id === id);
+        this._selectedUid.set(activeTab?.uId || undefined);
+        if (scroll && activeTab) {
+            this._scrollToPanel(activeTab);
+        }
     }
 
     /**
@@ -200,4 +360,90 @@ export class IconTabBarComponent implements OnInit {
         }
         return styles;
     }
+
+    private _generateTabConfig(tab: IconTabBarTabComponent): TabConfig {
+        const tabConfig: TabConfig = {
+            label: tab.label,
+            color: tab.color,
+            active: tab.active,
+            counter: tab.counter,
+            renderer: tab.renderer,
+            id: tab.id,
+            subItems: tab.children.map((c) => this._generateTabConfig(c))
+        };
+
+        return tabConfig;
+    }
+
+    /** @hidden */
+    private _scrollToPanel(tabPanel: Nullable<IconTabBarTabContentDirective>): void {
+        if (!tabPanel) {
+            return;
+        }
+        const panelElement = tabPanel.elementRef.nativeElement;
+        const containerElement = this._scrollbar.elementRef.nativeElement;
+        const distanceToScroll = panelElement.offsetTop - containerElement.offsetTop;
+        const maximumScrollTop = containerElement.scrollHeight - containerElement.clientHeight;
+        const currentScrollPosition = Math.ceil(containerElement.scrollTop);
+
+        if (!(currentScrollPosition === maximumScrollTop && distanceToScroll > maximumScrollTop)) {
+            this._disableScrollSpy = true;
+            fromEvent(containerElement, 'scroll')
+                .pipe(debounceTime(100), first(), takeUntilDestroyed(this._destroyRef))
+                .subscribe(() => (this._disableScrollSpy = false));
+            scrollTop(containerElement, distanceToScroll);
+        }
+    }
+
+    /**
+     * @hidden
+     * @description generate IconTabItems from TabConfig array
+     */
+    private _generateTabBarItems(
+        config: TabConfig[],
+        indexPrefix = '',
+        flatIndexRef?: FlatIndex,
+        parentUId?: string
+    ): IconTabBarItem[] {
+        flatIndexRef = flatIndexRef || { value: 0 };
+        return config.map((item, index) => {
+            const uId = `${indexPrefix}${index}`;
+            const result: IconTabBarItem = {
+                ...item,
+                index,
+                cssClasses: [],
+                uId,
+                hidden: false,
+                parentUId,
+                flatIndex: flatIndexRef!.value++,
+                subItems: this._generateTabBarItems(
+                    item.subItems || [],
+                    `${uId}${UNIQUE_KEY_SEPARATOR}`,
+                    flatIndexRef,
+                    uId
+                )
+            };
+            if (item.color) {
+                result.cssClasses = [`fd-icon-tab-bar__item--${item.color}`];
+            }
+            return result;
+        });
+    }
+
+    private _generateFlatTabs(tabs: IconTabBarItem[]): IconTabBarItem[] {
+        const flatTabs = tabs.reduce((acc: IconTabBarItem[], tab) => {
+            acc.push(tab);
+            if (tab.subItems) {
+                acc.push(...this._generateFlatTabs(tab.subItems));
+            }
+            return acc;
+        }, []);
+
+        return flatTabs;
+    }
+}
+
+/** @hidden helper object that is used in tab generation functions to calculate their indexes not depending on level of nesting */
+interface FlatIndex {
+    value: number;
 }