+++ libs/platform/icon-tab-bar/icon-tab-bar.module.ts
@@ -3,12 +3,14 @@ import { NgModule } from '@angular/core';
 import { IconTabBarFilterTypeComponent } from './components/icon-tab-bar-filter-type/icon-tab-bar-filter-type.component';
 import { IconTabBarIconTypeComponent } from './components/icon-tab-bar-icon-type/icon-tab-bar-icon-type.component';
 import { IconTabBarProcessTypeComponent } from './components/icon-tab-bar-process-type/icon-tab-bar-process-type.component';
+import { IconTabBarTabComponent } from './components/icon-tab-bar-tab/icon-tab-bar-tab.component';
 import { IconTabBarTextTypeComponent } from './components/icon-tab-bar-text-type/icon-tab-bar-text-type.component';
 import { IconTabBarPopoverComponent } from './components/popovers/icon-tab-bar-popover/icon-tab-bar-popover.component';
 import { TextTypePopoverComponent } from './components/popovers/text-type-popover/text-type-popover.component';
 import { IconBarDndContainerDirective } from './directives/dnd/icon-bar-dnd-container.directive';
 import { IconBarDndItemDirective } from './directives/dnd/icon-bar-dnd-item.directive';
 import { IconBarDndListDirective } from './directives/dnd/icon-bar-dnd-list.directive';
+import { IconTabBarFreeContentDirective } from './directives/icon-tab-bar-free-content.directive';
 import { IconTabBarComponent } from './icon-tab-bar.component';
 
 const components = [
@@ -21,8 +23,12 @@ const components = [
     TextTypePopoverComponent,
     IconBarDndListDirective,
     IconBarDndItemDirective,
-    IconBarDndContainerDirective
+    IconBarDndContainerDirective,
+    IconTabBarTabComponent,
+    IconTabBarFreeContentDirective
 ];
+
+export const FDP_ICON_TAB_BAR = [IconTabBarComponent, IconTabBarTabComponent, IconTabBarFreeContentDirective] as const;
 
 /**
  * @deprecated